add replacement for configomat.sh
I wrote a function that ShellCheck does not want to immediately kill
(like `configomat.sh` ..). The new function is written in idiomatic Bash
and properly documented. The behavior is nearly identical to the old
script; the only exception is that you cannot provide more than one file
to adjust contents in. This could be easily done in the future;
currently we don't need it and it would make the function longer than
necessary (I like KISS).