small misc adjustments
I noticed that shutting down the whole container in case the file you
want to change stuff in does not exist might not be a good idea after
all. I discovered this because it seems that `/etc/postfix/maps/sender_login_maps.ldap`
is not actually present when running the tests. I added an extra check.
We did not discover this previously because it was hidden by the usage
of `$(...)`, in which the call to `configomat.sh` with a file that
doesnt exist was the last call. The script would call exit, which would
exit the subshell spawned by `$(...)`. We didn't notice it because we do
not use `set -eE`. The new (and proper) function reveiled this flaw.