fix message about dir not existing (OpenDKIM)
For Rspamd, the test showed an error which stated that Bash could not
write to `/etc/opendkim/trustedHosts`. The directory is not created as
the test disables OpenDKIM & OpenDMARC. Rspamd setup can run without
them as Rspamd can do DKIM/DMARC itself.

I added a proper guard before the `echo` lines that were causing the
error to appear.