fixed small bug for values with spaces
So, because commands like `set-option-for-controller` only usually take
2 arguments, previously the value (argument 2) was broken in case it
contained a space (as the `while read` loop splits at whitespaces). The
other part of value after the space would go into argument 3. I
corrected this behaviour. Now, we pass in argument 3 as well and cut off
the last whitespace if it exists.