completely refactor README & parts of docs
I removed many parts of the `README` and placed them inside the docs.
This was done to ensure we can properly version all of out
documentation. I believe that users should generally **always use our
documentation**, not a mixture of `README` + docs, to configure and
check their mail server. This makes maintenance easier for maintainers
as well as changes are local to the version. Previously, we had to keep
`README` and docs in sync, which is undesirable.

I took many parts of the READMe and copy-pasted it into the docs. Other
parts were slightly adjusted to match the context. The landing page of
our docs was more heavily refactored and updated to display more
information.

I also added notes on the versioning of the docs!

There is actually more work to do cleaning up parts of the docs (a
section on the examples chapter needs overhaul, etc.). But I did want to
make this PR bigger - it is big enough.