adjust `_send_mail` for the last time :D
CONTAINER_NAME is now assumed to be set, which makes invocation even
easier for common cases. Moreover, the helper that additionally
retrieves the mail ID was renamed (_send_mail_* -> _send_email_*) to be
canonical with _send_mail; it also takes the same arguments now, and has
the same assumptions when calling it.