cleanup
Remove functionality from `start-mailserver.sh` to put it into the
appropriate stacks; this way, `start-mailserver.sh` does not need to
know about any arrays at all and the arrays specific for each stack are
in each stack. The functions for registering were moved as well, since
this nicely encapsulates all state for a stack now in the stack file.