moved setup functionality into setup stack
The `start-daemons.sh` stack contained setup functionality, which I
moved into the setup stack (for Postgrey, whose removal of PID files was
not necessary because we already do that in `_setup_apply_fixes_after_configuration`)
& Dovevot, for which the code for starting the daemon contained plain
setup functionality entirely unrelated to the actual start).

Sorry for requiring a review for this again, but I did not want to open
yet another PR for this (since I'm currently swamping the project with
PRs). Should be fearly easy to review this commit though, so all is well
:)