improve Postfix config
This commit fixes #3117 as it improves Postfix's configuration
to only add `reject_authenticated_sender_login` to services/ports where
SASL is used (getting rid of the warning messages shown in the issue).

The wording for the log was chosen as we need to be consistent with the
default value, which is `0`.