rename `ENABLE_REDIS` -> `ENABLE_RSPAMD_REDIS`
I know, this is a breaking change, but it's fine since the var has not
actually been in a release yet.

We use Redis only for Rspamd, so I want to "namespace" it. I want to
provide another feature that should be namespaced to Rspamd as well.
Hence, all is fitting.