improve current Rspamd & Dovecot Sieve tests
Added a few more checks to the tests and added a custom
`user-patches.sh` to simplify setup. This will come in handy later.

The Sieve tests were adjusted because I had an eye on them as I will
adjust the Sieve code with the next commit and I saw that
they could use some of our nifty helpers.