tests: Change OpenLDAP image to `bitnami/openldap`
`osixia/openldap` has not seen any activity since Feb 2021, while our `Dockerfile` was fixed to v1.1.6` (Feb 2018).

This commit migrates to `bitnami/openldap` which required modifying the `01_mail-tree.ldif` to also include adding the root object to start successfully. Additionally the image does not support the `.schema` format, they must be converted to `.ldif` which has been done for `postfix-book.schema`.

The user account `.ldif` files have minimal changes.
- Lines moved around for better organization
- Additional comments for context
- Removal of inherited `objectClass` attributes (`person`, `top`) from the `orgnizationalPerson` class. Attribute `sn` changed to long form `surname` and values corrected with `givenName`. `changetype: add` was also not necessary.

We no longer need to build an image with the LDIF files to use it. This image is actively maintained and one of the most popular OpenLDAP images on DockerHub.

Startup time is around 5 seconds? The LDAP test uses a standard 20 second timeout check to wait until the server is ready before continuing with starting DMS.