ci: Simplify `.ecrs.json` config
Better align with the `.editorconfig` files rules (_which `eclint` reads_).

Most of this is the upstream defaults and can be dropped.
- `SpaceAftertabs` (_possibly invalid name?_) was set to `true` but it doesn't appear any files covered rely on this. No context in `git blame` for it's addition, dropping.
- `Exclude` has a long history of pattern additions with minimal context. This is updated to be more specific now.
- `IgnoreDefaults` has been left for visibility. It could technically be set to `true`, but the SVG files should be ignored then.

`test/test-files/` is only ignored due to an upstream bug with `eclint` incorrectly detecting a file with the wrong encoding and triggering a failure as it's unsupported by `eclint`. Can be removed once a new release resolves that.

`test/bats/` and it's related submodules `bats-assert` / `bats-support` need to be ignored. In addition to not being relevant to the project, they supply their own `.editorconfig` with rules that seem to fail such as `max_line_length`.

`.git/` also isn't ignored by default (_it apparently was in an earlier version of `eclint` according to a bug report_).