ci: Adjust `.editorconfig`
We can bring the `indent_size = 2` rule into all files default. It helps establish a bit of consistency in the project files.

`.gitmodules` could possibly be ignored instead, it's also using tabs like `Makefile` so I've just configured that explicitly here.

The git submodules section doesn't seem relevant? Those changes shouldn't be committed, and we're not likely to modify those intentionally. The project supplies it's own `.editorconfig` file as root which would ignore this section, thus it's not really serving a purpose?