fix: Ensure `.svbin` files are newer than `.sieve` source files
This change prevents operations applied to both files that adjusts their `mtime` attribute which sometimes results in the same timestamp which causes Dovecot to log errors.

Functionally there doesn't appear to be any issue. After logging the error, the existing svbin is used and works correctly.

This does help avoid false-positives in the test-suite that are asserting for errors in a log.