docs: Introduce `TaskList` for pro/con lists
Bullet points aren't that nice when we're differentiating with an icon already, but `mkdocs-material` makes it difficult to target via CSS as we cannot easily add a list to a `ul` element and it'd be a bit awkward to append to each list item.

Task list feature works well, but customizing the unchecked style would look better. This is acceptable since we're not a TaskList anywhere else.